Created a shared testing engine
