+++ projects/rudder-controller/project/Tests/utest.h
@@ -1,0 +1,19 @@
+/*
+ * unit_tests.h
+ *
+ *  Created on: Feb 24, 2024
+ *      Author: Sailbot
+ */
+
+#ifndef SRC_UNIT_TESTS_H_
+#define SRC_UNIT_TESTS_H_
+
+#include "test_engine.h"
+
+
+void run_tests(void);
+
+// Test functions
+testresult read_bearing_test(void);
+
+#endif /* SRC_UNIT_TESTS_H_ */
