+++ projects/template/project/Core/Src/main.c
@@ -18,7 +18,7 @@
 /* USER CODE END Header */
 /* Includes ------------------------------------------------------------------*/
 #include "main.h"
-#include "unit_tests.h"
+#include "utest.h"
 /* Private includes ----------------------------------------------------------*/
 /* USER CODE BEGIN Includes */
 
@@ -85,7 +85,7 @@ int main(void)
   /* Initialize all configured peripherals */
   /* USER CODE BEGIN 2 */
 #ifdef TEST_MODE
-	  test_main();
+  run_tests();
 #endif
   /* USER CODE END 2 */
