+++ projects/template/project/Debug/sources.mk
@@ -26,4 +26,5 @@ Core/Src \
 Core/Startup \
 Drivers/STM32U5xx_HAL_Driver/Src \
 Tests \
+test_framework \
