+++ projects/template/project/Tests/uconfig.h
@@ -1,0 +1,18 @@
+/*
+ * test_config.h
+ *
+ *  Created on: Feb 27, 2024
+ *      Author: mlokh
+ */
+
+#ifndef UCONFIG_H_
+#define UCONFIG_H_
+
+typedef enum {
+	ALL,
+	TEMP
+} testgroup;
+
+extern const testgroup TEST_GROUP_SEL;
+
+#endif /* UCONFIG_H_ */
