+++ projects/template/project/Tests/utest.c
@@ -1,0 +1,39 @@
+/*
+ * utest.c
+ *
+ *  Created on: Feb 27, 2024
+ *      Author: mlokh
+ */
+#include "utest.h"
+
+#define NUM_ITERS 10
+
+const testgroup TEST_GROUP_SEL = ALL;
+
+// -- Add to test runner here --
+const t_test test_runner[] = {
+//		{"Name of test", "function definition", "testgroup id"
+		{.testname="Temporary test 1", .func=temp_test1, .group=TEMP},
+		{.testname="Iteration test", .func=iteration_test}
+};
+
+
+// -- Unit tests --
+testresult temp_test1(void) {
+	testresult res = {TSUCCESS, {0}};
+	return res;
+}
+
+testresult iteration_test(void) {
+	testresult res = {TSUCCESS, {0}};
+
+	for (int i = 0; i < NUM_ITERS; ++i) {
+		// res.error.raw = someFunction(..);
+		HAL_Delay(1000);
+	}
+	return res;
+}
+
+void run_tests(void) {
+	test_main(test_runner, sizeof(test_runner) / sizeof(t_test));
+}
