+++ projects/template/project/Tests/utest.h
@@ -1,0 +1,20 @@
+/*
+ * utest.h
+ *
+ *  Created on: Feb 27, 2024
+ *      Author: mlokh
+ */
+
+#ifndef UTEST_H_
+#define UTEST_H_
+
+#include "test_engine.h"
+
+
+void run_tests(void);
+
+// Test functions
+testresult temp_test1(void);
+testresult iteration_test(void);
+
+#endif /* UTEST_H_ */
